requirements.txt
flask>=3.0.0
flask-login>=0.6.3
selenium>=4.0.0
pandas>=2.0.0
fpdf2>=2.7.0
python-docx>=1.0.0
openpyxl>=3.1.0
requests>=2.31.0
cryptography>=42.0.0
python-dotenv>=1.0.0
